requirements.txt
Pillow
pyautogui
pywin32
requests
